+++ src/main/kotlin/io/snyk/plugin/services/SnykTaskQueueService.kt
@@ -68,14 +68,6 @@ class SnykTaskQueueService(val project: Project) {
     @TestOnly
     fun getTaskQueue() = taskQueue
 
-    fun scheduleRunnable(title: String, runnable: (indicator: ProgressIndicator) -> Unit) {
-        taskQueue.run(object : Task.Backgroundable(project, title, true) {
-            override fun run(indicator: ProgressIndicator) {
-                runnable.invoke(indicator)
-            }
-        })
-    }
-
     fun connectProjectToLanguageServer(project: Project) {
             // subscribe to the settings changed topic
         val languageServerWrapper = LanguageServerWrapper.getInstance()
@@ -114,7 +106,7 @@ class SnykTaskQueueService(val project: Project) {
                 waitUntilCliDownloadedIfNeeded()
                 indicator.checkCanceled()
 
-                if (settings.snykCodeSecurityIssuesScanEnable || settings.snykCodeQualityIssuesScanEnable) {
+                if (settings.snykCodeSecurityIssuesScanEnable || settings.snykCodeQualityIssuesScanEnable || isSnykOSSLSEnabled()) {
                     if (!isStartup) {
                         LanguageServerWrapper.getInstance().sendScanCommand(project)
                     }
@@ -122,12 +114,6 @@ class SnykTaskQueueService(val project: Project) {
                 if (settings.ossScanEnable) {
                     if (!isSnykOSSLSEnabled()) {
                         scheduleOssScan()
-                    } else {
-                        // the LS deals with triggering scans at startup
-                        // TODO: Refactor when more than Snyk Code is available in LS for IntelliJ
-                        if (!isStartup) {
-                            LanguageServerWrapper.getInstance().sendScanCommand(project)
-                        }
                     }
                 }
                 if (isIacEnabled() && settings.iacScanEnabled) {