+++ src/main/kotlin/io/snyk/plugin/ui/SnykSettingsDialog.kt
@@ -85,9 +85,8 @@ class SnykSettingsDialog(
     private val scanOnSaveCheckbox = JCheckBox().apply { toolTipText = "If enabled, automatically scan on save, start-up and configuration change" }
     private val additionalParametersTextField: JTextField = ExpandableTextField().apply { toolTipText = "--all-projects is already defaulted, -d causes problems" }
 
-    private val scanTypesPanelOuter = ScanTypesPanel(project, rootPanel)
-    private val codeAlertPanel = scanTypesPanelOuter.codeAlertPanel
-    private val scanTypesPanel = scanTypesPanelOuter.panel
+    private val scanTypesPanelOuter = ScanTypesPanel(project)
+    private val scanTypesPanel = scanTypesPanelOuter.scanTypesPanel
 
     private val issueViewOptionsPanel = IssueViewOptionsPanel(project).panel
 
@@ -149,7 +148,6 @@ class SnykSettingsDialog(
             cliBaseDownloadUrlTextField.text = applicationSettings.cliBaseDownloadURL
             additionalParametersTextField.text = applicationSettings.getAdditionalParameters(project)
             scanOnSaveCheckbox.isSelected = applicationSettings.scanOnSave
-
             cliReleaseChannelDropDown.selectedItem = applicationSettings.cliReleaseChannel
         }
     }
@@ -332,7 +330,7 @@ class SnykSettingsDialog(
                 )
             )
         }
-        val productAndSeveritiesPanel = JPanel(UIGridLayoutManager(6, 4, JBUI.emptyInsets(), 30, -1))
+        val productAndSeveritiesPanel = JPanel(UIGridLayoutManager(1, 2, JBUI.emptyInsets(), 30, -1))
 
         rootPanel.add(
             productAndSeveritiesPanel,
@@ -345,28 +343,13 @@ class SnykSettingsDialog(
             )
         )
 
-        val productSelectionPanel = JPanel(UIGridLayoutManager(5, 4, JBUI.emptyInsets(), -1, -1))
-        productSelectionPanel.border = IdeBorderFactory.createTitledBorder("Product selection")
-
         productAndSeveritiesPanel.add(
-            productSelectionPanel,
+            scanTypesPanel,
             baseGridConstraints(
                 row = 0,
                 anchor = UIGridConstraints.ANCHOR_NORTHWEST,
                 fill = UIGridConstraints.FILL_HORIZONTAL,
                 hSizePolicy = UIGridConstraints.SIZEPOLICY_CAN_SHRINK or UIGridConstraints.SIZEPOLICY_CAN_GROW,
-                indent = 0
-            )
-        )
-
-        productSelectionPanel.add(
-            scanTypesPanel,
-            baseGridConstraints(
-                row = 0,
-                anchor = UIGridConstraints.ANCHOR_NORTHWEST,
-                fill = UIGridConstraints.FILL_NONE,
-                hSizePolicy = UIGridConstraints.SIZEPOLICY_CAN_SHRINK or UIGridConstraints.SIZEPOLICY_CAN_GROW,
-                vSizePolicy = UIGridConstraints.SIZEPOLICY_CAN_SHRINK or UIGridConstraints.SIZEPOLICY_CAN_GROW,
                 indent = 0
             )
         )
@@ -395,15 +378,6 @@ class SnykSettingsDialog(
                 hSizePolicy = UIGridConstraints.SIZEPOLICY_CAN_SHRINK or UIGridConstraints.SIZEPOLICY_CAN_GROW,
                 vSizePolicy = UIGridConstraints.SIZEPOLICY_CAN_SHRINK or UIGridConstraints.SIZEPOLICY_CAN_GROW,
                 indent = 0
-            )
-        )
-
-        rootPanel.add(
-            codeAlertPanel,
-            baseGridConstraints(
-                row = 2,
-                anchor = UIGridConstraints.ANCHOR_NORTHWEST,
-                indent = 2
             )
         )