+++ src/main/kotlin/io/snyk/plugin/ui/toolwindow/SnykToolWindowFactory.kt
@@ -5,7 +5,6 @@ import com.intellij.openapi.project.Project
 import com.intellij.openapi.util.Disposer
 import com.intellij.openapi.wm.ToolWindow
 import com.intellij.openapi.wm.ToolWindowFactory
-import io.snyk.plugin.pluginSettings
 
 /**
  * IntelliJ ToolWindowFactory for Snyk plugin.