+++ src/main/kotlin/snyk/common/SnykCachedResults.kt
@@ -1,6 +1,5 @@
 package snyk.common
 
-import com.google.gson.Gson
 import com.intellij.openapi.Disposable
 import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.components.Service
@@ -167,7 +166,7 @@ class SnykCachedResults(
                             currentSnykCodeError =
                                 SnykError(
                                     snykScan.cliError?.error ?: snykScan.errorMessage
-                                        ?: "Failed to run Snyk Code Scan",
+                                    ?: "Failed to run Snyk Code Scan",
                                     snykScan.cliError?.path ?: snykScan.folderPath,
                                     snykScan.cliError?.code,
                                 )
@@ -196,7 +195,7 @@ class SnykCachedResults(
 
                     SnykBalloonNotificationHelper
                         .showError(
-                            "scanning error for project ${project.name}. Data: ${snykScan}",
+                            "scanning error for project ${project.name}. Data: $snykScan",
                             project,
                         )
                 }
@@ -220,7 +219,7 @@ class SnykCachedResults(
 
                         }
 
-                        LsProductConstants.Container.value-> {
+                        LsProductConstants.Container.value -> {
 
                         }
                     }
@@ -230,6 +229,7 @@ class SnykCachedResults(
     }
 }
 
+// TODO use this!
 internal class SnykFileIssueComparator(
     private val snykResults: Map<SnykFile, List<ScanIssue>>,
 ) : Comparator<SnykFile> {