+++ src/test/kotlin/io/snyk/plugin/ui/toolwindow/SnykToolWindowPanelTest.kt
@@ -10,27 +10,28 @@ import io.mockk.justRun
 import io.mockk.mockk
 import io.mockk.unmockkAll
 import io.mockk.verify
-import io.snyk.plugin.net.CliConfigSettings
-import io.snyk.plugin.net.LocalCodeEngine
 import io.snyk.plugin.pluginSettings
-import io.snyk.plugin.services.SnykAnalyticsService
-import io.snyk.plugin.services.SnykApiService
 import io.snyk.plugin.services.SnykApplicationSettingsStateService
 import io.snyk.plugin.services.SnykTaskQueueService
+import org.eclipse.lsp4j.services.LanguageServer
+import org.eclipse.lsp4j.services.WorkspaceService
 import org.junit.Test
 import snyk.UIComponentFinder
-import snyk.amplitude.AmplitudeExperimentService
+import snyk.common.lsp.LanguageServerWrapper
+import snyk.common.lsp.SnykLanguageClient
 import snyk.oss.OssVulnerabilitiesForFile
 import java.awt.Container
 import java.io.File
+import java.util.concurrent.CompletableFuture
 
 class SnykToolWindowPanelTest : LightPlatform4TestCase() {
-    private val analyticsService: SnykAnalyticsService = mockk(relaxed = true)
     private val taskQueueService = mockk<SnykTaskQueueService>(relaxed = true)
-    private val snykApiServiceMock = mockk<SnykApiService>()
     private val settings = mockk<SnykApplicationSettingsStateService>(relaxed = true)
-    private val amplitudeExperimentationServiceMock = mockk<AmplitudeExperimentService>()
     private lateinit var cut: SnykToolWindowPanel
+    val lsMock = mockk<LanguageServer>()
+    val lsClientMock = mockk<SnykLanguageClient>()
+    val lsProcessMock = mockk<Process>()
+    val workspaceServiceMock = mockk<WorkspaceService>()
 
     override fun setUp() {
         super.setUp()
@@ -38,31 +39,47 @@ class SnykToolWindowPanelTest : LightPlatform4TestCase() {
 
         val application = ApplicationManager.getApplication()
         application.replaceService(SnykApplicationSettingsStateService::class.java, settings, application)
-        application.replaceService(SnykApiService::class.java, snykApiServiceMock, application)
-        application.replaceService(SnykAnalyticsService::class.java, analyticsService, application)
-        application.replaceService(
-            AmplitudeExperimentService::class.java,
-            amplitudeExperimentationServiceMock,
-            application
-        )
 
         project.replaceService(SnykTaskQueueService::class.java, taskQueueService, project)
+
+        val lsw = LanguageServerWrapper.getInstance()
+        lsw.languageServer = lsMock
+        lsw.languageClient = lsClientMock
+        lsw.process = lsProcessMock
+
+        every { lsProcessMock.info().startInstant().isPresent } returns true
+        every { lsProcessMock.isAlive } returns true
+        every { lsMock.workspaceService } returns workspaceServiceMock
+        val sastSettings = mapOf(
+            Pair("sastEnabled", true),
+            Pair(
+                "localCodeEngine", mapOf(
+                    Pair("allowCloudUpload", false),
+                    Pair("enabled", false),
+                    Pair("url", "")
+                )
+            ),
+            Pair("org", "1234"),
+            Pair("reportFalsePositivesEnabled", false),
+            Pair("autofixEnabled", false),
+            Pair("supportedLanguages", emptyList<String>()),
+        )
+        every { workspaceServiceMock.executeCommand(any()) } returns CompletableFuture.completedFuture(sastSettings)
 
         every { settings.token } returns null
         every { settings.sastOnServerEnabled } returns true
         every { settings.localCodeEngineEnabled } returns false
-        every { snykApiServiceMock.getSastSettings()?.sastEnabled } returns true
-        every { snykApiServiceMock.getSastSettings()?.localCodeEngine?.enabled } returns false
     }
 
     override fun tearDown() {
         unmockkAll()
 
         val application = ApplicationManager.getApplication()
-        application.replaceService(AmplitudeExperimentService::class.java, AmplitudeExperimentService(), application)
-        application.replaceService(SnykApplicationSettingsStateService::class.java, SnykApplicationSettingsStateService(), application)
-        application.replaceService(SnykApiService::class.java, SnykApiService(), application)
-        application.replaceService(SnykAnalyticsService::class.java, SnykAnalyticsService(), application)
+        application.replaceService(
+            SnykApplicationSettingsStateService::class.java,
+            SnykApplicationSettingsStateService(),
+            application
+        )
 
         project.replaceService(SnykTaskQueueService::class.java, SnykTaskQueueService(project), project)
         super.tearDown()
@@ -122,10 +139,6 @@ class SnykToolWindowPanelTest : LightPlatform4TestCase() {
     fun `should not display onboarding panel and run scan directly`() {
         every { settings.token } returns "test-token"
         every { settings.pluginFirstRun } returns true
-        every { snykApiServiceMock.getSastSettings() } returns CliConfigSettings(
-            true,
-            LocalCodeEngine(false, "", false),
-        )
         justRun { taskQueueService.scan(false) }
 
         cut = SnykToolWindowPanel(project)
@@ -136,15 +149,10 @@ class SnykToolWindowPanelTest : LightPlatform4TestCase() {
         assertEquals(findOnePixelSplitter(vulnerabilityTree), descriptionPanel!!.parent)
 
         verify(exactly = 1) { taskQueueService.scan(false) }
-        verify(exactly = 1) { analyticsService.logAnalysisIsTriggered(any()) }
     }
 
     @Test
     fun `should automatically enable all products on first run after Auth`() {
-        every { snykApiServiceMock.getSastSettings() } returns CliConfigSettings(
-            true,
-            LocalCodeEngine(false, "", false),
-        )
         val application = ApplicationManager.getApplication()
         application.replaceService(
             SnykApplicationSettingsStateService::class.java,
@@ -156,9 +164,6 @@ class SnykToolWindowPanelTest : LightPlatform4TestCase() {
 
         SnykToolWindowPanel(project)
 
-        verify(exactly = 1, timeout = 5000) {
-            snykApiServiceMock.getSastSettings()
-        }
         assertTrue(pluginSettings().ossScanEnable)
         assertTrue(pluginSettings().snykCodeSecurityIssuesScanEnable)
         assertTrue(pluginSettings().snykCodeQualityIssuesScanEnable)
@@ -168,10 +173,6 @@ class SnykToolWindowPanelTest : LightPlatform4TestCase() {
 
     @Test
     fun `should automatically enable all products on first run after Auth, with local engine enabled`() {
-        every { snykApiServiceMock.getSastSettings() } returns CliConfigSettings(
-            true,
-            LocalCodeEngine(true, "http://foo.bar/api", false),
-        )
         val application = ApplicationManager.getApplication()
         application.replaceService(
             SnykApplicationSettingsStateService::class.java,
@@ -181,11 +182,10 @@ class SnykToolWindowPanelTest : LightPlatform4TestCase() {
         pluginSettings().token = "test-token"
         pluginSettings().pluginFirstRun = true
 
+        assertTrue(LanguageServerWrapper.getInstance().isInitialized)
+
         SnykToolWindowPanel(project)
 
-        verify(exactly = 1, timeout = 5000) {
-            snykApiServiceMock.getSastSettings()
-        }
         assertTrue(pluginSettings().ossScanEnable)
         assertTrue(pluginSettings().snykCodeSecurityIssuesScanEnable)
         assertTrue(pluginSettings().snykCodeQualityIssuesScanEnable)